LoginTest

+++ src/test/java/testBases/BaseClass.java
@@ -1,15 +1,25 @@
 package testBases;
 
+import java.io.File;
+
 import java.io.FileReader;
 import java.io.IOException;
 import java.net.URL;
+import java.text.SimpleDateFormat;
 import java.time.Duration;
+import java.util.Date;
 import java.util.Properties;
 
 import org.apache.commons.lang3.RandomStringUtils;
+import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.openqa.selenium.OutputType;
 import org.openqa.selenium.Platform;
+import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.annotations.AfterClass;
@@ -28,10 +38,11 @@ public class BaseClass {
 	public void setup(String os, String br) throws IOException
 	{
 		//Config.properties file
-		FileReader file=new FileReader(".src//test//resources//config.properties");
+		FileReader file=new FileReader("./src//test//resources//config.properties");
 		p=new Properties();
 		p.load(file);
 		
+		logger =LogManager.getLogger(this.getClass());
 		
 		//OS and Browser
 		if(p.getProperty("execution_envir").equalsIgnoreCase("remote"))
@@ -71,6 +82,17 @@ public class BaseClass {
 					
 		}
 		
+		if(p.getProperty("execution_envir").equalsIgnoreCase("local"));
+		{
+			switch(br.toLowerCase())
+			{
+			case "chrome" : driver=new ChromeDriver(); break;
+			case "firefox" : driver=new FirefoxDriver(); break;
+			case "edge" : driver=new EdgeDriver(); break;
+			default : System.out.println("invalid browser name..."); return;
+			}
+		}
+		
 		
 		driver.manage().deleteAllCookies();
 		driver.manage().window().maximize();
@@ -84,9 +106,9 @@ public class BaseClass {
 	}
 	
 	@AfterClass
-	public void teardown() throws InterruptedException
+	public void teardown() 
 	{
-		Thread.sleep(3000);
+		
 		driver.close();
 	}
 	
@@ -102,11 +124,26 @@ public class BaseClass {
 		return generatednum;
 	}
 	
-	public String alphanumber()
+	public String alphanumeric()
 	{
 		String randomname=RandomStringUtils.randomAlphabetic(3);
 		String randomnum=RandomStringUtils.randomAlphanumeric(4);
 		return(randomname+"!@"+randomnum);
+	}
+	
+	public String captureScreen(String tname) throws IOException
+	{
+		String timeStamp = new SimpleDateFormat("yyyyMMddHHmmss").format(new Date());
+		
+		TakesScreenshot takesScreenshot =(TakesScreenshot) driver; 
+		File sourceFile = takesScreenshot.getScreenshotAs(OutputType.FILE);
+		
+		String targetFilePath=System.getProperty("user.dir")+"\\screenshots\\" + tname + "_" + timeStamp + ".png";
+		File targetFile=new File(targetFilePath);
+		
+		sourceFile.renameTo(targetFile);
+		return targetFilePath;
+		
 	}
 
 }
